wh4-load1

+++ repository/contacts.js
@@ -1,33 +1,33 @@
 const Contact = require('../models/contact')
 
-const listContacts = async () => {
-  const result = await Contact.find()
+const listContacts = async (query, user) => {
+  const result = await Contact.find({owner: user.id})
   return result
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({_id: contactId})
+const getContactById = async (contactId, user) => {
+  const result = await Contact.findOne({_id: contactId, owner: user.id}).populate({path:'owner', select: 'name email role createdAt updatedAt'})
   return result
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({_id: contactId})
+const removeContact = async (contactId, user) => {
+  const result = await Contact.findOneAndRemove({_id: contactId, owner: user.id})
   return result
 }
 
-const addContact = async (body) => {
-  const result = await Contact.create(body)
+const addContact = async (body, user) => {
+  const result = await Contact.create({...body, owner: user.id})
   return result
 }
 
-const updateContact = async (contactId, body) => {
-  const result = await Contact.findOneAndUpdate({_id: contactId}, {...body}, {new: true})
+const updateContact = async (contactId, body, user) => {
+  const result = await Contact.findOneAndUpdate({_id: contactId, owner: user.id}, {...body}, {new: true})
   return result
 }
 
-const updateFavorite = async (contactId, body) => {
+const updateFavorite = async (contactId, body, user) => {
     const result = await Contact
-    .findByIdAndUpdate({_id: contactId}, {...body}, {new: true})
+    .findByIdAndUpdate({_id: contactId, owner: user.id}, {...body}, {new: true})
     return result
   }
